+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
   "editor.codeActionsOnSave": {
     "source.organizeImports": true,
-    "source.fixAll": true
+    "source.fixAll": true,
+    "source.addMissingImports": true
   }
 }
