+++ src/api/dto/comment.ts
@@ -1,0 +1,11 @@
+import { UploaderDTO } from "./user";
+
+export interface CommentDTO {
+  id: number;
+  body: string;
+  author: UploaderDTO;
+  imageId: number;
+  parentId?: number;
+  replies?: CommentDTO[];
+  replies_count?: number;
+}
