+++ src/api/dto/image.ts
@@ -1,24 +1,7 @@
-export interface TagDTO {
-  label: string;
-  value: string;
-  id?: string;
-  name?: string;
-}
+import { CommentDTO } from "./comment";
+import { TagDTO } from "./tag";
+import { UploaderDTO } from "./user";
 
-export interface UploaderDTO {
-  id: number;
-  username: string;
-}
-
-export interface CommentDTO {
-  id: number;
-  body: string;
-  author: UploaderDTO;
-  imageId: number;
-  parentId?: number;
-  replies?: CommentDTO[];
-  replies_count?: number;
-}
 export default interface ImageDTO {
   id?: number;
   caption?: string;
@@ -29,4 +12,19 @@ export default interface ImageDTO {
   comments?: CommentDTO[];
   likes_count?: number;
   created_at?: Date;
+}
+
+export interface PostedImageDTO {
+  id: number;
+  caption: string;
+  url: string;
+  thumbnailUrl: string;
+  uploader: UploaderDTO;
+  tags: TagDTO[] | [];
+  comments: CommentDTO[] | [];
+  likes_count: number;
+  created_at: Date;
+  formatedCreatedOnDate: string;
+  isLiked: boolean;
+  isSaved: boolean;
 }
