+++ src/api/dto/tag.ts
@@ -1,0 +1,6 @@
+export interface TagDTO {
+  label: string;
+  value: string;
+  id?: string;
+  name?: string;
+}
