+++ src/api/services/post-services.ts
@@ -1,40 +1,39 @@
 import { formatDistanceToNow } from "date-fns";
+import { interpolate } from "src/utils/string";
 import { jelloWithAuth } from "..";
-import { CommentDTO, TagDTO, UploaderDTO } from "../dto/image";
+import { PostedImageDTO } from "../dto/image";
 import endpoints from "../endpoints";
 
-export interface PostedImageDTO {
-  id: number;
-  caption: string;
-  url: string;
-  thumbnailUrl: string;
-  uploader: UploaderDTO;
-  tags: TagDTO[] | [];
-  comments: CommentDTO[] | [];
-  likes_count: number;
-  created_at: Date;
-  formatedCreatedOnDate: string;
-}
+const fromJSON = (img: PostedImageDTO): PostedImageDTO => {
+  const formattedDate = new Date(img.created_at);
 
-const getFeedPosts = async (): Promise<PostedImageDTO[]> => {
-  return jelloWithAuth.get(endpoints.POPULAR).then((res) => {
-    return res?.data.map((img: PostedImageDTO) => {
-      const formattedDate = new Date(img.created_at);
-
-      return {
-        ...img,
-        formatedCreatedOnDate: formatDistanceToNow(formattedDate),
-      };
-    });
-  });
+  return {
+    ...img,
+    formatedCreatedOnDate: formatDistanceToNow(formattedDate),
+  };
 };
 
+const getFeedPosts = async (): Promise<PostedImageDTO[]> => {
+  return jelloWithAuth
+    .get(endpoints.POPULAR)
+    .then((res) => res?.data.map(fromJSON));
+};
+
+const likePost = async (postId: number) => {
+  return jelloWithAuth.post(interpolate(endpoints.LIKE_IMAGE, { postId }));
+};
+
+const unLikePost = async (postId: number) => {
+  return jelloWithAuth.post(interpolate(endpoints.LIKE_IMAGE, { postId }));
+};
 // const getPersonalPosts = async () => {
 
 // }
 
 const PostService = {
   getFeedPosts,
+  likePost,
+  unLikePost,
 };
 
 export default PostService;
